Keep dropping and creating array_accum()
We want to keep array_accum() around for now to support any custom SQL that
sites might have generated. Plus, we don't want it to show up in our search
path ahead of the native array_agg()!

We can always rebase this as a fix into the previous commit to squash history a
bit. "Nope, that never happened!"

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>