Use copy owning library as the context for finding best hold sort order setting
In the olden days before best hold sort order and proximity adjustment, there was
no need to consider any context org other than "here", because the choices were,
essentially, FIFO or Traditional applied within the hard-boundary range.  However,
now we need to decided whose rule to follow, and the most correct arbiter of that
is the owning library of the call number of the copy.  We avoid copy circ lib
because that can change with floating copies.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Ben Shum <bshum@biblio.org>