LP1271218 - clarify how opt-in user notifications are shown in my account
As reported in LP127128, end users may not need to be shown the message that
"No notification preferences are configured" when in fact this message is only
displayed when the user opt-in notification settings are not configured.

So, alter this to only show the table and div when there are settings to be
configured and otherwise show nothing.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Chris Sharp <csharp@georgialibraries.org>
Signed-off-by: Ben Shum <bshum@biblio.org>