KPAC library page: give the kids some style
Just the bare minimum: some padding around the body, and a font colour and size
that's reasonably consistent with the rest of the look of the KPAC for h2 and
h3 elements.

Yes, I know there is text-transform:uppercase and h2 is sized at 48px but that
seems to be used for "You saved it!" and "Get it!" banner things. On the
library info page, we're using headings for headings.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>