LP1223903 - Rewrite marc_export.in in support-scripts.
This commit rewrites marc_export.in in the support-scripts  src directory.
It will still be transformed into marc_export at build time.  This change
replaces the current marc_export scipt with a faster, DBI based alternative.

The replacement support nearly all of the features of the original, except
for the progress output.  Much of this information is harder to gather
with the new design.

This is a squashed, rebased commit of previous work.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Ben Shum <bshum@biblio.org>