LP1125471 - Use label instead of id in recording in-house use prompt
When recording multiple in-house use with the library settings for threshold
prompts, the prompt would bring up the ID as the variable instead of the name
of the non-cataloged material. This caused confusion among staff members who
might see a message like: "Are you sure you want to mark 9 as having been used
21 times?" (9 being the id, not the name of the non-cataloged material)

This commit leverages existing code which had been commented out and under
the ID to use the label name for the non-cataloged material instead. Also,
minor bugfix with a missing semi-colon at the end of the function being used.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Bill Erickson <berick@esilibrary.com>