LP#1053397 filter MR icons/formats repair
Only limit the scope of the MR attributes in the presence of a hard hold
boundary when searching for which attributes to display in the hold
placement form.

In other words, don't limit the attributes based on search parameters,
since searching at Branch 1 does not necesssarily meany you only want to
place holds on formats available at branch 1.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Dan Wells <dbw2@calvin.edu>