LP#1282704: Serials: in routing lists, avoid extra page breaks wasting paper at print time
If you're using batch receive, and you have two copies (and therefore
two streams) that you're receiving with the Routing List? checkbox
checked, but only one of those streams actually has a routing list,
you'll note before this patch is applied that you get an extra page of
blank paper if you print the routing list that pops up.

This fixes that.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>