LP#800478 repair logic error in acq fund transfer
The source data for the amount applied to each allocation and transfer
(in certain circumstances) was incorrect.

The transfer amount should not be the total amount allocated from the
source to the fund, but instead the amount of the funding source credit
in cases where we have to pull money from multiple funding source
credits.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Liam Whalen <liam.whalen@bc.libraries.coop>
Signed-off-by: Mike Rylander <mrylander@gmail.com>