LP#1286248: remove references to osrf_ctl.sh
This patch replaces references to osrf_ctl.sh, which is deprecated
as of OpenSRF 2.3.0, with opensrf_control.  It also names
2.3.0 as the minimum required version of OpenSRF.

brick_ctl.sh gets the most changes as a consequence.  In particular:

* the stop_{c,perl}/[re]start_{c,perl} actions are removed, as
  osrf_control behaves exactly the same when stopping or starting
  services.
* the -s option for working with a single service is now documented,
  and is restricted to the stop_osrf/[re]start_osrf actions

It should be noted that for brick-based setups (and other kinds of
multi-server setups) better control can be had by using opensrf_control
directly.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Dan Wells <dbw2@calvin.edu>