LP#1284864 MR opac-visible constituent records
The number of constituent records for a metarecord during searching is
now consitently based on the presence of an opac visible copy and not a
holdable copy.  This affects the constituent record counts in MR results
lists and the records displayed in the constituent records list.  (Icons
were already checking visibility instead of holdability).

Holdable formats are still derived from records with holdable copies.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>