LP#1287791: Restrict authority browse to controlled subfields
Back in bug 1175308 I added some logic that improves the behavior
of the MARC editor's right-click authority popup. Now we need to
go a little farther and adjust that logic to be more careful about
which subfields it uses to browse for authority records. Specifically,
it should only use controlled subfields.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Remington Steed <rjs7@calvin.edu>
Signed-off-by: Dan Wells <dbw2@calvin.edu>