LP#1077811 Inconsistent INSTALL Instructions
The preamble to the installation instructions, intended for
developers, is a bit unclear. The instructions describe actions
that should take place at two different points in the installation.

This commit moves the instructions to download and install the Dojo
Toolkit until after the opensrf Linux user already owns the
/openils directory. It also adds the appropriate note for non-devs
to skip this section.

Signed-off-by: Elliot V <evoris@slcconline.edu>
Signed-off-by: Remington Steed <rjs7@calvin.edu>