LP#1259665 Split series browse config from search
When we added browse support in 2.5, we took extra care to ensure that
our series browse entries would sort and display reliably.

800t does not specify non-filing characters, so it was excluded from
browse.  Unfortunately, since browse and search had a single root
config, this removed 800t from search as well.

While we may want to reconsider expanding later on, this commit takes
a conservative approach, and uses all series values for search and
facets, but only entries with a usable filing indicator will be
extracted for browse.  It accomplishes this by splitting the series
config in two, one config for search/faceting, and a second config
dedicated to browse (similar to what was done for title).

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Ben Shum <bshum@biblio.org>