LP#1303987: Fully populate Ills from context menu
The Ills fixed field is 4 characters long (allowing for up to four
types of illustrations to be noted, per 008/006), but the fixed field
editor dropdown menus are not (a) multi-select capable nor (b) ready
to support multi-value-per-field, because, well Ills is the only one
like this, and that would be a pretty big complication to the code.
However, even though multi-select is not supported, we should still
apply the value selected and update the fixed field, otherwise the
value is not updated in the record itself.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Remington Steed <rjs7@calvin.edu>
Signed-off-by: Dan Wells <dbw2@calvin.edu>