LP#1304559 Fix slow Vandelay-based imports
Whether a planning limitation or otherwise, we have seen a few cases
where queries run very, very slowly (appear to hang) when using the
updated metabib.record_attr view in a WHERE clause.

The simplest code change is to switch Vandelay to use the intermediate
'_flat' view instead, and early reports indicate that this at least
gets us back to a functional state.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Ben Shum <bshum@biblio.org>