LP#1306133: Avoid metabib.rec_descriptor where possible
metabib.rec_descriptor has been a view over other data structures
for some time now, but with the advent of the new attribute infrastructure
it is a view over a view over a view, variously transforming IDs to
a slim table to an HSTORE structure and finally to a wide table. Thus,
it is extremely expensive to query. So we won't unless we're asked to.
Which, in the TPAC world, is never.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Doug Kyle <dkyle@grpl.org>