LP#1306176: force installation of Business::Stripe from CPAN
Business::Stripe can fail to pass its tests on platforms that
happen to have Pod::Simple 3.24 or later and Test::Pod.  In
particular, Debian Jessie and Ubuntu Trusty meet those conditions
(since the OpenSRF installation process installs libtest-pod-perl)
and Debian Wheezy with backports enabled does as well.

Since the failing test cases are only for POD, this patches moves
Business::Stripe to the CPAN_MODULES_FORCE list across the board.

This can be undone if/when Business::Stripe releases a fix for
the POD validation warnings.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Bill Erickson <berick@esilibrary.com>