Docs 2.6: Lost Item Billing Min/Max Price Settings
Add a short description for the new 2.6 feature "Lost Item Billing: New
Min/Max Price Settings" to the only section where it seems to fit, and it
seems obscure and hard to find. It could be useful to create a new, more
prominent section for common Org Unit Settings, possibly organized by
category. Sending this via GitHub to test the new Docs workflow.

Signed-off-by: Remington Steed <rjs7@calvin.edu>