LP#1314370: Enable dojo all the time in the catalog
Because we rely on the dojo framework for the shelving location filter in the
advanced search, we should choose to use dojo everywhere and not just when
autosuggest, google books, or novelist are enabled.

This is a quick change to toggle the flag for want_dojo to be set all the time.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Liam Whalen <liam.whalen@bc.libraries.coop>