LP#1312945: auth-auth linking: cache less agressively and look for all links
There was a logic inversion that was causing miss-use of a per-
record cache mechanism.  I remove that entirely, as it's not helpful
in practice.

Also, after finding one use of a linked field we moved on to the
next, though it is possible for more than one link to exist for
different uses of a name, say.  Now we'll loop through all fields
to find all possible linkages.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Srey Seng <sreyseng@gmail.com>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>