LP#1312945: authority.calculate_authority_linking and multiple linked tags
The function "authority.calculate_authority_linking" was not able to deal with
situations where there are multiple marc tags that are linkable. The function
only process the first tag that contains a linking_subfield and returns that
to be inserted into the authority_linking table, even when there are more than
one of the same tag with linking_subfield.

Added an additional loop to loop through each tag, for situations where there are
multiples of each tag.

Signed-off-by: Srey Seng <sreyseng@gmail.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>