LP#925776: Recheck located uri visibility
Previous to this commit, once we have determined that there are no
local Located URIs or foreign copies, we include the record in the
result set in the staff client.  However, the desire of library staff
is that foreign Located URIs be valid as a visibility limiter.  So
we now recheck at the end for that situation.  This honors the new
"Located URIs act like copies" global flag.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Ben Shum <bshum@biblio.org>