LP#1310751: Reorder query-building wrapper for QueryParser
In order to make sure that the transparent boolean pushdown does
not cause filters and modifiers to end up deeply embedded in the
query tree, we put them at the front.

This is a short-term fix.  Long-term, we need to make use of the
floating subquery syntax ( {{ ... }} ) available in QueryParser,
which is designed to address the semantic separation between what
the user types into the search box, and the various UI widgets that
can add query clauses.

Signed-off-by: Jason Etheridge <jason@esilibrary.com>
Signed-off-by: Dan Wells <dbw2@calvin.edu>