LP#1325720 Repair fund rollover year selector query
Dojo dojo stores track values as strings, so values used in data store
queries must be coerced into strings before performing the query.

This fixes the year query in the fund rollover UI so that the page is
able to render the results of a rollover without dying from an JS error
before completing.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Ben Shum <bshum@biblio.org>