LP#1326149 Use a TPAC-settable TIME_FORMAT for local time formats
We had tried to use the format.time library setting, but that is meant
for Dojo which uses the Unicode formats, that are incompatible with
the POSIX strftime() formats required by Template::Plugin::Date.

At the same time, we need to expose the raw machine-readable time
for schema.org consumers, so express that in a @content attribute.

And we might as well use a more meaningful <time> element instead of
the meaningless <span>, just for good measure.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>