LP#1326983: excluded fulfilled holds when adding hold_request.shelf_expires_soon events
This patch adjusts the example A/T filter for the
hold_request.shelf_expires_soon hook to exclude hold requests that are
already marked as fulfilled.  This saves time creating events that
would immediately be marked as invalid per the HoldIsAvailable
validator.

It also addresses an issue for large databases that have lots of
fulfilled holds wherein hold_request.shelf_expires_soon events can
fail to be added at all due to a cstore timeout.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Ben Shum <bshum@biblio.org>