LP1198475: Add a new, optional, Lost and Paid copy status.
Add the Lost and Paid status to config.copy_status.

Add org_unit_settig to control if the copy status is used.

Add checks to O::A::Circ::Money::make_payments to check for LOST
status and org_unit setting and change the copy status to
Lost and Paid as appropriate.

Also add checks for Lost and Paid copy status wherever LOST copy
status is used.

Add the COPY_STATUS_LOST_AND_PAID event.

Add the event to ils_events.xml with id 7026.

Check for the event in circ/util.js in the staff client.

Add pgTAP tests for the existence of the Lost and Paid copy status
and the new org_unit setting to control if it isused.

Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Rogan Hamby <rogan.hamby@gmail.com>