LP#968514: Revive bib/auth importing docs from 2.1, add intro
This commit revives the following sections from the 2.1 version of the docs
and updates them for 2.6:

  - marc_export: Exporting Bibliographic Records into MARC files
    - includes new options for 2.6
  - Importing Authority Records from Command Line
  - Importing Authority Records from the Staff Client

The first two are included in a new section called "Support Scripts"
within the "Developer Resources" section, with a new intro and summary
of other commonly used scripts provided with Evergreen. The third section
is moved inside the Cataloging chapter "Batch Importing MARC Records",
with slight modifications to the intro to that chapter.

Signed-off-by: Remington Steed <rjs7@calvin.edu>