LP#1331127 Repair sort logic of previous issuances
Ensure that the list of previous issuances is sorted correctly (on
date_published) when looking for the previous serial.unit to update its
copy location (when serial.prev_issuance_copy_location is enabled).

The data comes sorted from the DB, but the sorting was lost during the
process of unique-ifying the list.  Ultimately, it was relying on the
order of hash keys, which is undefined.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Dan Wells <dbw2@calvin.edu>