LP#1322285: Exclude default values for fixed fields when ingesting
Default values for fixed fields that can exist in either 008 or 006
will cause search to be over-broad.  So, we avoid including those now.

I also took the time to move the baseline schema around a bit in order
to keep related code in more appropriate files and avoid continuously
redefining functions.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>